Did some MISRA compliance correction (part 3)

+++ middlewares/libpus/inc/pus_tools/tables_management.h
@@ -29,7 +29,7 @@
  * @def  BUILD_ROUTING_KEY(apid, service, subservice)
  * @brief Preprocessor function that build routing key with APID, sevice and subservice
  */
-#define BUILD_ROUTING_KEY(apid, service, subservice)    ((uint32_t)((apid << 16) | (service << 8) | subservice))
+#define BUILD_ROUTING_KEY(apid, service, subservice)    ((uint32_t)(((uint32_t)(apid) << 16) | ((uint32_t)(service) << 8) | (uint32_t)(subservice)))
 
 /***************************** Types Definitions *****************************/
